refactor: 도메인을 세부적인 디렉토리로 구분
