+++ src/main/java/domain/card/CardShuffler.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public interface CardShuffler {
     int getRandomCardIndex(int deckSize);
