+++ src/main/java/domain/result/DealerMatchResult.java
@@ -1,0 +1,4 @@
+package domain.result;
+
+public record DealerMatchResult(int winCount, int drawCount, int loseCount) {
+}
