+++ src/main/java/domain/result/MatchCase.java
@@ -1,0 +1,5 @@
+package domain.result;
+
+public enum MatchCase {
+    WIN, DRAW, LOSE
+}
