+++ src/test/java/domain/card/DeckTest.java
@@ -1,11 +1,9 @@
-package domain;
+package domain.card;
 
-import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
