+++ src/test/java/domain/pariticipant/NameTest.java
@@ -1,6 +1,5 @@
-package domain;
+package domain.pariticipant;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
